Populate initial bubblesort hierarchy
Create basic structure for the first project `Sorting and Searching',
first milestone `Bubble Sort'.

Implements `make_random_array' but actually spell it as
`MakeRandomArray' because in Go it is preferred to use MixedCaps or
mixedCaps (please see `Effective Go'!).

For the moment let's put it in its own `utils' package and also write
some basic tests for it to asserts what was requested in the
milestone.